+++ lite3d/include/lite3d/lite3d_mesh_loader.h
@@ -36,36 +36,36 @@ LITE3D_CEXPORT int lite3d_mesh_load_from_m_file(lite3d_mesh *mesh, const lite3d_
 
 LITE3D_CEXPORT int lite3d_mesh_indexed_load_from_memory(lite3d_mesh *mesh, 
     const void *vertices, 
-    size_t verticesCount, 
+    uint32_t verticesCount,
     const lite3d_mesh_layout *layout,
-    size_t layoutCount,
+    uint32_t layoutCount,
     const void *indexes, 
-    size_t elementsCount, 
+    uint32_t elementsCount,
     uint8_t indexComponentSize, 
     uint16_t access);
 
 LITE3D_CEXPORT int lite3d_mesh_indexed_extend_from_memory(lite3d_mesh *mesh, 
     const void *vertices, 
-    size_t verticesCount, 
+    uint32_t verticesCount,
     const lite3d_mesh_layout *layout,
-    size_t layoutCount,
+    uint32_t layoutCount,
     const void *indexes, 
-    size_t elementsCount, 
+    uint32_t elementsCount,
     uint8_t indexComponentSize, 
     uint16_t access);
 
 LITE3D_CEXPORT int lite3d_mesh_load_from_memory(lite3d_mesh *mesh, 
     const void *vertices, 
-    size_t verticesCount, 
+    uint32_t verticesCount,
     const lite3d_mesh_layout *layout,
-    size_t layoutCount,
+    uint32_t layoutCount,
     uint16_t access);
 
 LITE3D_CEXPORT int lite3d_mesh_extend_from_memory(lite3d_mesh *mesh, 
     const void *vertices, 
-    size_t verticesCount, 
+    uint32_t verticesCount,
     const lite3d_mesh_layout *layout,
-    size_t layoutCount,
+    uint32_t layoutCount,
     uint16_t access);
 
 LITE3D_CEXPORT void lite3d_mesh_order_mat_indexes(lite3d_mesh *mesh);
